update (#1)
* fix export csv

* fix lead capture

* fix mention hook

* fix portal route

* fix formula int parsing

* fix formula int parsing

* lead capture tool

* get rid of phone and email attribute types

* field converer class name in metadata

* fix warning

* css changes

* fix dashboard template

* update gridstack

* css fix

* loader improvement

* preload gridstack

* update bull

* refactoring

* dashboard fixes

* css change

* css changes

* rebuild fixes

* emailaddress phonenumber repository fixes

* cs fix

* array value: use transactions

* cs fix

* cs fixes and array values cleanup

* cron manager fix

* fix class parser

* cs fix

* comments fix

* DataCache util

* fix test

* cs fix

* change metadata loader

* using data cache

* var export fix

* amend

* data cache usage

* data cache usage

* data cache usage

* data cache usage

* fix varexport

* Data Cache usage

* data cache usage

* data cache usage

* fix tests

* cs fix

* imap: support tls

* fix email account

* fix upgrade

* fix bc

* email account fixes

* bc fix

* upgrade fix

* upgrade command fixes

* package lock version

* cleanup

* fix language

* Upgrade fixes

* cs fix

* cs fixes

* fix record grid

* v

* fix portal list view url

* cs fix

* entry point route detection

* portal index refactoring

* fix dashlet layout field

* fix portal role clear cache

* cs fix

* fix tester

* email: only read access in poral roles

* fix field manager

* varchar 150 max length by default

* fix notice

* Added new options in upgrade command

* fix

* portal fixes

* cs fixes

* cs fix

* json retrieve function

* fix docs link

* email folder select modal css fix

* email: move to folder from trash

* fix json function

* fix email import

* ldap fix

* cleanup

* ldap fix

* cs fix

* ldap fix

* v

* cs fix

* fix notice

* file rename in loop for windows

* cs fix

* data cache support onlt array and stdclass

* fix ldap

* data cache get safe contents

* fix autoload

* upgrade real estate compatibility check

* orm fix order list

* orm cleanup

* list tree css fix

* settings tooltips

* image preview improvements

* categories tree improvements

* css fix

* fix admin panel

* v

* cs fix

* cs fix

* cs fix

* calendar modal fix

* cs fix

* fix image orientation

* fix image orientation

* date after before validation autocomplete

* getPhpSafeContents for unix

* Installation fixes

* cs fix

* fix convert lead

* fix password recovery

* Installation fixes

* portal action list restriction

* Upgrade script

* Upgrade script

* Upgrade script

* fix email modal

* upgrade cleanup files and fix msg

* fix diff

* fix diff: ignore upgrades

* v

* upgrade cleanup

* label fixes

* fix label

* fix reminder submition

* fix import button

* application state changes

* hook fix

* email auto reply error log

* fix htmlizer and email template no id

* cs fix

* ouauth callback changes

* entry point changes

* select manager fix

* remove self assign in portal

* fix request getContentType

* fix export

* fix email invalid on bounced

* fix campaign log search

* campaign log search by email address

* search queue items

* email queue items filters email

* export fixes

* fix error msg

* fix mass actions

* cs fix

* Fix

* Fix

* Bug fixes

* fix upgrade script

* v

* case close reject actions only if available in status list

* kanban refactoring

* fix language manager

* fix export

* v

* Update Admin.json (#1814)

* Update Admin.json (#1814)

* formula fix logical not

* formula fix logical operator priority

* fix export all fields

* v

* diff all patch versions

* fix 2fa

* fix export specific fields

* fix activities and select manager

* fix tester

* address view remove duplicated codes (#1823)

* v

* hide email menu items in portal

* fix email folder

* fix pdf page number

* template manager image

* add log to slim

* cs fix

* popup notifications grouped request

* kanban order

* fix notifications

* fix portal runner

* v

* currency format 10 $

* export xlsx currency format

* cleanup and cs fix

* fix link search blur

* Upgrade fix

* search apply button

* entity manager page

* Command fix

* fix text filter by foreign field

* auth token manager

* fix task panel layout

* repository: calling updateFetchedValues always after save

* fix acl check command

* preferences fixes

* app service changes

* fix export

* update npm packages

* update packages

* fix person name

* Fixed PHP binary path

* Fixed PHP binary path

* fix duplicate attachments multiple

* v

* di binding

* fix typo

* app info improvements

* fix export currency

* v

* fix deprecated

* fix deprecated

+++ application/Espo/Core/Upgrades/Actions/Base.php
@@ -33,7 +33,10 @@ use Espo\Core\Utils\Util;
 use Espo\Core\Utils\System;
 use Espo\Core\Utils\Json;
 use Espo\Core\Exceptions\Error;
+
 use Composer\Semver\Semver;
+
+use Throwable;
 
 abstract class Base
 {
@@ -52,6 +55,8 @@ abstract class Base
     private $databaseHelper;
 
     protected $processId = null;
+
+    protected $parentProcessId = null;
 
     protected $manifestName = 'manifest.json';
 
@@ -152,11 +157,19 @@ abstract class Base
         return $this->getContainer()->get('entityManager');
     }
 
-    public function throwErrorAndRemovePackage($errorMessage = '')
+    public function throwErrorAndRemovePackage($errorMessage = '', $deletePackage = true, $systemRebuild = true)
     {
-        $this->deletePackageFiles();
-        $this->deletePackageArchive();
-        $this->disableMaintenanceMode();
+        if ($deletePackage) {
+            $this->deletePackageFiles();
+            $this->deletePackageArchive();
+        }
+
+        $this->disableMaintenanceMode(true);
+
+        if ($systemRebuild) {
+            $this->systemRebuild();
+        }
+
         throw new Error($errorMessage);
     }
 
@@ -182,9 +195,19 @@ abstract class Base
         return $this->processId;
     }
 
+    protected function getParentProcessId()
+    {
+        return $this->parentProcessId;
+    }
+
     public function setProcessId($processId)
     {
         $this->processId = $processId;
+    }
+
+    public function setParentProcessId($processId)
+    {
+        $this->parentProcessId = $processId;
     }
 
     /**
@@ -243,9 +266,11 @@ abstract class Base
 
         foreach ($versionList as $version) {
             $isInRange = false;
+
             try {
                 $isInRange = Semver::satisfies($currentVersion, $version);
-            } catch (\Throwable $e) {
+            }
+            catch (Throwable $e) {
                 $GLOBALS['log']->error('SemVer: Version identification error: '.$e->getMessage().'.');
             }
 
@@ -309,11 +334,13 @@ abstract class Base
             $scriptName = $scriptNames[$type];
 
             require_once($beforeInstallScript);
+
             $script = new $scriptName();
 
             try {
                 $script->run($this->getContainer(), $this->scriptParams);
-            } catch (\Throwable $e) {
+            }
+            catch (Throwable $e) {
                 $this->throwErrorAndRemovePackage($e->getMessage());
             }
         }
@@ -494,7 +521,8 @@ abstract class Base
     {
         try {
             $res = $this->getFileManager()->copy($sourcePath, $destPath, $recursively, $fileList, $copyOnlyFiles);
-        } catch (\Throwable $e) {
+        }
+        catch (Throwable $e) {
             $this->throwErrorAndRemovePackage($e->getMessage());
         }
 
@@ -667,12 +695,12 @@ abstract class Base
         $packageArchivePath = $this->getPackagePath(true);
 
         if (!file_exists($packageArchivePath)) {
-            throw new Error('Package Archive doesn\'t exist.');
+            $this->throwErrorAndRemovePackage('Package Archive doesn\'t exist.', false, false);
         }
 
         $res = $this->getZipUtil()->unzip($packageArchivePath, $packagePath);
         if ($res === false) {
-            throw new Error('Unnable to unzip the file - '.$packagePath.'.');
+            $this->throwErrorAndRemovePackage('Unnable to unzip the file - '.$packagePath.'.', false, false);
         }
     }
 
@@ -705,9 +733,12 @@ abstract class Base
     protected function systemRebuild()
     {
         try {
-            return $this->getContainer()->get('dataManager')->rebuild();
-        } catch (\Throwable $e) {
-            $GLOBALS['log']->error('Database rebuild failure, details: '.$e->getMessage().'.');
+            $this->getContainer()->get('dataManager')->rebuild();
+
+            return true;
+        }
+        catch (Throwable $e) {
+            $GLOBALS['log']->error('Database rebuild failure, details: '. $e->getMessage() .'.');
         }
 
         return false;
@@ -768,7 +799,7 @@ abstract class Base
             $permissionDeniedList = $this->getFileManager()->getLastPermissionDeniedList();
 
             $delimiter = $this->isCli() ? "\n" : "<br>";
-            throw new Error("Permission denied: " . $delimiter . implode($delimiter, $permissionDeniedList));
+            $this->throwErrorAndRemovePackage("Permission denied: " . $delimiter . implode($delimiter, $permissionDeniedList), false, false);
         }
     }
 
@@ -806,6 +837,12 @@ abstract class Base
     protected function enableMaintenanceMode()
     {
         $config = $this->getConfig();
+        $configParamName = $this->getTemporaryConfigParamName();
+        $parentConfigParamName = $this->getTemporaryConfigParamName(true);
+
+        if ($config->has($configParamName) || ($parentConfigParamName && $config->has($parentConfigParamName))) {
+            return;
+        }
 
         $actualParams = [
             'maintenanceMode' => $config->get('maintenanceMode'),
@@ -813,7 +850,6 @@ abstract class Base
             'useCache' => $config->get('useCache'),
         ];
 
-        $configParamName = $this->getTemporaryConfigParamName();
         $config->set($configParamName, $actualParams);
 
         $save = false;
@@ -838,23 +874,32 @@ abstract class Base
         }
     }
 
-    protected function disableMaintenanceMode()
+    protected function disableMaintenanceMode($force = false)
     {
         $config = $this->getConfig();
 
-        $configParamName = $this->getTemporaryConfigParamName();
-        $temporaryUpgradeParams = $config->get($configParamName, []);
+        $configParamList = [
+            $this->getTemporaryConfigParamName()
+        ];
+
+        if ($force && $this->getTemporaryConfigParamName(true)) {
+            $configParamList[] = $this->getTemporaryConfigParamName(true);
+        }
 
         $save = false;
 
-        foreach ($temporaryUpgradeParams as $paramName => $paramValue) {
-            if ($config->get($paramName) != $paramValue) {
-                $config->set($paramName, $paramValue);
-                $save = true;
-            }
-        }
+        foreach ($configParamList as $configParamName) {
 
-        if ($config->has($configParamName)) {
+            if (!$config->has($configParamName)) {
+                continue;
+            }
+
+            foreach ($config->get($configParamName, []) as $paramName => $paramValue) {
+                if ($config->get($paramName) != $paramValue) {
+                    $config->set($paramName, $paramValue);
+                }
+            }
+
             $config->remove($configParamName);
             $save = true;
         }
@@ -864,9 +909,18 @@ abstract class Base
         }
     }
 
-    protected function getTemporaryConfigParamName()
+    protected function getTemporaryConfigParamName($isParentProcess = false)
     {
-        return 'temporaryUpgradeParams' . $this->getProcessId();
+        $processId = $this->getProcessId();
+
+        if ($isParentProcess) {
+            $processId = $this->getParentProcessId();
+            if (!$processId) {
+                return;
+            }
+        }
+
+        return 'temporaryUpgradeParams' . $processId;
     }
 
     protected function isCli()
